+++ CHANGELOG.md
@@ -2,3 +2,4 @@
   
 
 - [world] feature: now only sends the `dimension created success` and `dimension deleted success` messages to the `command source`. (Instead of `broadcast`)
+- [world] fix: the console spam while deleting a dimension.